add toutiao model save and imageinfo save

+++ FrameMobile/src/FrameMobile.Model/News/NewsImageInfo.cs
@@ -10,7 +10,7 @@ namespace FrameMobile.Model.News
     [SubSonicTableNameOverride("NewsImageInfo")]
     public class NewsImageInfo : MySQLModelBase
     {
-        public int NewsId { get; set; }
+        public long NewsId { get; set; }
 
         public int Width { get; set; }
